directory restructure for reducers and graphs directory. Continue working on tests
