+++ public/js/audio.js
@@ -1,4 +1,3 @@
-import { configureStore } from '@reduxjs/toolkit';
 import DOM from './dom.js';
 import Reverb from './reverb.js';
 import Distortion from './distortion.js';
@@ -6,8 +5,6 @@ import Gain from './gain.js';
 import createTrack from './track.js';
 import waveformVisualizer from './visualizer.js';
 import impulseResponse from './impulse-response.js';
-import graphReducer from './reducer.js';
-import node from './node.js';
 
 const createGraph = (nodes, track) =>
   nodes.reduce((acc, node) => acc.connect(node), track);
@@ -28,15 +25,6 @@ export default async function initialize() {
   const preAmpGainNode = Gain(audioContext);
   const postAmpGainNode = Gain(audioContext, { gain: 2 });
   const reverbNode = await Reverb(audioContext);
-
-  const store = configureStore({ reducer: graphReducer });
-  console.log(postAmpGainNode, convolver);
-  store.dispatch({ type: 'connect', node: node(preAmpGainNode, 0) });
-  store.dispatch({ type: 'connect', node: node(postAmpGainNode, 1) });
-  store.dispatch({ type: 'connect', node: node(reverbNode, 2) });
-  console.log(store.getState());
-  store.dispatch({ type: 'disconnect', pos: 0 });
-  console.log(store.getState());
 
   const nodes = [
     preAmpGainNode,
