+++ public/js/graph/index.js
@@ -1,0 +1,3 @@
+import audioGraphNode from './node';
+
+export { audioGraphNode };
