+++ public/js/graph/node.js
@@ -1,0 +1,8 @@
+function audioGraphNode(node, pos) {
+  return {
+    pos,
+    type: node?.name,
+  };
+}
+
+export default audioGraphNode;
