--- public/js/node.js
@@ -1,8 +1,0 @@
-function node(audioNode, pos) {
-  return {
-    pos,
-    type: audioNode?.name,
-  };
-}
-
-export default node;
