+++ public/js/reducers/graphReducer.js
