+++ public/js/reducers/index.js
@@ -1,0 +1,3 @@
+import graphReducer from './graphReducer';
+
+export { graphReducer };
